api/requirements.txt
asgiref==3.8.1
astroid==3.1.0
dill==0.3.8
Django==5.0.4
djangorestframework==3.15.1
isort==5.13.2
mccabe==0.7.0
platformdirs==4.2.1
psycopg2-binary==2.9.9
pylint==3.1.0
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
python-dotenv==1.0.1
sqlparse==0.5.0
tomli==2.0.1
tomlkit==0.12.4
typing_extensions==4.11.0
